+++ src/main/java/com/example/stagemate/domain/performance/PerformanceType.java
@@ -1,4 +1,4 @@
-package com.example.stagemate.domain.performances;
+package com.example.stagemate.domain.performance;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
